Implemented extra test cases
- Implemented operations on structs

- Added plenty of test cases to try to break it

- Cleaned up some files

+++ compiler_tests/extratests/extra_assign/mixed_assignment_driver.c
@@ -1,4 +1,4 @@
-int f();
+double f();
 
 int main()
 {
